+++ inst/examples/ex-seek.R
@@ -1,6 +1,10 @@
 ## Seek columns
-seek_columns(iris, select = startsWith, prefix = "Sepal")
+seek_columns(iris, select = startsWith, prefix = "Petal")
+seek_columns(iris, names = c("Petal.Length", "Petal.Width"))
 
 ## Get columns
-x <- get_columns(iris, select = startsWith, prefix = "Sepal")
+x <- get_columns(iris, select = startsWith, prefix = "Petal")
+head(x)
+
+x <- get_columns(iris, names = c("Petal.Length", "Petal.Width"))
 head(x)